v1.0 for whomever wants it. Is no longer found anywhere so I decided to post it on my GIT.
I was not and am not developing this. I might in the future take a look at the code but I simply was asked to make it available because I had it and someone wants it. So use at your own risk. I use to run this but run a different version now.